Code save. Input names are set according to the number of added inputs
Previous version was not setting the input names and resetting the form
which results in lots of errors once the form was being submitted. This
version tries to fix that although other errors appears. This commit is
meant to be a code save and future development will have to continue
from this point onward.